fix(agent-runtime): move createAgentCard factory to AgentCardProperties
AgentCards (engine.spi) importing AgentCardProperties (runtime.boot) violated
the package boundary rule enforced by RuntimePackageBoundaryTest. Move the
properties-driven card factory into AgentCardProperties itself, so the
dependency flows boot → spi (allowed) rather than spi → boot (forbidden).
