+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/boot/AgentCardProperties.java
@@ -1,5 +1,11 @@
 package com.huawei.ascend.runtime.boot;
 
+import java.util.List;
+import org.a2aproject.sdk.spec.AgentCapabilities;
+import org.a2aproject.sdk.spec.AgentCard;
+import org.a2aproject.sdk.spec.AgentInterface;
+import org.a2aproject.sdk.spec.AgentProvider;
+import org.a2aproject.sdk.spec.TransportProtocol;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 /**
@@ -63,5 +69,42 @@ public class AgentCardProperties {
     /** Returns {@code true} when the user explicitly configured a card name. */
     boolean hasExplicitName() {
         return name != null && !name.isBlank();
+    }
+
+    /**
+     * Build an {@link AgentCard} using this properties object's values for
+     * any set fields, with sensible defaults for unset fields.
+     *
+     * @param name the card name (non-blank; supplied by the caller)
+     */
+    public AgentCard createAgentCard(String name) {
+        String resolvedDescription = blankToDefault(description, "agent-runtime");
+        String resolvedVersion = blankToDefault(version, "0.1.0");
+        String resolvedEndpoint = blankToDefault(endpoint, "/a2a");
+        String resolvedOrg = blankToDefault(organization, "spring-ai-ascend");
+        String resolvedOrgUrl = blankToDefault(organizationUrl, "http://localhost:8080");
+
+        AgentCapabilities capabilities = AgentCapabilities.builder()
+                .streaming(true)
+                .pushNotifications(true)
+                .extendedAgentCard(false)
+                .build();
+        return AgentCard.builder()
+                .name(name)
+                .description(resolvedDescription)
+                .url(resolvedEndpoint)
+                .version(resolvedVersion)
+                .provider(new AgentProvider(resolvedOrg, resolvedOrgUrl))
+                .capabilities(capabilities)
+                .defaultInputModes(List.of("text"))
+                .defaultOutputModes(List.of("text", "artifact"))
+                .skills(List.of())
+                .supportedInterfaces(List.of(new AgentInterface(TransportProtocol.JSONRPC.asString(), resolvedEndpoint)))
+                .preferredTransport(TransportProtocol.JSONRPC.asString())
+                .build();
+    }
+
+    private static String blankToDefault(String value, String defaultValue) {
+        return value == null || value.isBlank() ? defaultValue : value;
     }
 }
