+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/engine/spi/AgentCards.java
@@ -1,7 +1,6 @@
 package com.huawei.ascend.runtime.engine.spi;
 
 import java.util.List;
-import com.huawei.ascend.runtime.boot.AgentCardProperties;
 import org.a2aproject.sdk.spec.AgentCapabilities;
 import org.a2aproject.sdk.spec.AgentCard;
 import org.a2aproject.sdk.spec.AgentInterface;
@@ -43,41 +42,5 @@ public final class AgentCards {
                 .supportedInterfaces(List.of(new AgentInterface(TransportProtocol.JSONRPC.asString(), endpoint)))
                 .preferredTransport(TransportProtocol.JSONRPC.asString())
                 .build();
-    }
-
-    /**
-     * Build an {@link AgentCard} from YAML-driven properties, using sensible
-     * defaults for any unset fields. The caller is responsible for providing
-     * a non-blank {@code name}.
-     */
-    public static AgentCard createFromProperties(String name, AgentCardProperties properties) {
-        String description = blankToDefault(properties.getDescription(), "agent-runtime");
-        String version = blankToDefault(properties.getVersion(), "0.1.0");
-        String endpoint = blankToDefault(properties.getEndpoint(), "/a2a");
-        String organization = blankToDefault(properties.getOrganization(), "spring-ai-ascend");
-        String organizationUrl = blankToDefault(properties.getOrganizationUrl(), "http://localhost:8080");
-
-        AgentCapabilities capabilities = AgentCapabilities.builder()
-                .streaming(true)
-                .pushNotifications(true)
-                .extendedAgentCard(false)
-                .build();
-        return AgentCard.builder()
-                .name(name)
-                .description(description)
-                .url(endpoint)
-                .version(version)
-                .provider(new AgentProvider(organization, organizationUrl))
-                .capabilities(capabilities)
-                .defaultInputModes(List.of("text"))
-                .defaultOutputModes(List.of("text", "artifact"))
-                .skills(List.of())
-                .supportedInterfaces(List.of(new AgentInterface(TransportProtocol.JSONRPC.asString(), endpoint)))
-                .preferredTransport(TransportProtocol.JSONRPC.asString())
-                .build();
-    }
-
-    private static String blankToDefault(String value, String defaultValue) {
-        return value == null || value.isBlank() ? defaultValue : value;
     }
 }
